requirements.txt
Django==1.9.6
django-crispy-forms==1.6.1
Pillow


